+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## v1.0.3-dev
+## v1.1.0-dev
+
+* Enhancements
+  * Optimize Tracker CRDT operations for increased performance
+  * Shard tracker internally to use pool of trackers for increased performance under load
+  * [Tracker] Add `get_by_key/3` to lookup a single presence entry for a given topic and key
 
 ## v1.0.2 (2017-06-14)